+++ packages/vuikit/src/library/navbar/components/navbar--logo-link.js
@@ -1,0 +1,7 @@
+import { ElementNavbarLogoLink } from '../elements'
+
+export default {
+  name: 'VkNavbarLogo',
+  functional: true,
+  render: ElementNavbarLogoLink.render
+}