+++ packages/vuikit/src/library/navbar/elements/navbar--logo-link.js
@@ -3,7 +3,7 @@ import { mergeData } from 'vuikit/src/util/vue'
 export default {
   functional: true,
   render (h, { data, children }) {
-    return h('span', mergeData(data, {
+    return h('a', mergeData(data, {
       class: 'uk-navbar-item uk-logo'
     }), children)
   }